+++ tests/Pest.php
@@ -1,5 +1,6 @@
 <?php
 
 use Cjmellor\Blockade\Tests\TestCase;
+use Illuminate\Foundation\Testing\RefreshDatabase;
 
-uses(TestCase::class)->in(__DIR__);
+uses(TestCase::class, RefreshDatabase::class)->in(__DIR__);
